refactor(app): clean up FillWell recovery options namespacing
