+++ app/src/organisms/ErrorRecoveryFlows/RecoveryOptions/RetryNewTips.tsx
@@ -1,13 +1,5 @@
-import { Trans, useTranslation } from 'react-i18next'
-
-import { LegacyStyledText } from '@opentrons/components'
-
 import { RECOVERY_MAP } from '../constants'
-import {
-  TwoColLwInfoAndDeck,
-  SelectTips,
-  TwoColTextAndFailedStepNextStep,
-} from '../shared'
+import { TwoColLwInfoAndDeck, SelectTips, RetryWithNewTips } from '../shared'
 import { SelectRecoveryOption } from './SelectRecoveryOption'
 
 import type { RecoveryContentProps } from '../types'
@@ -43,42 +35,4 @@ export function RetryNewTips(props: RecoveryContentProps): JSX.Element {
   }
 
   return buildContent()
-}
-
-export function RetryWithNewTips(props: RecoveryContentProps): JSX.Element {
-  const { recoveryCommands, routeUpdateActions } = props
-  const { retryFailedCommand, resumeRun } = recoveryCommands
-  const { handleMotionRouting } = routeUpdateActions
-  const { ROBOT_RETRYING_STEP } = RECOVERY_MAP
-  const { t } = useTranslation('error_recovery')
-
-  const primaryBtnOnClick = (): Promise<void> => {
-    return handleMotionRouting(true, ROBOT_RETRYING_STEP.ROUTE)
-      .then(() => retryFailedCommand())
-      .then(() => {
-        resumeRun()
-      })
-  }
-
-  const buildBodyText = (): JSX.Element => {
-    return (
-      <Trans
-        t={t}
-        i18nKey="robot_will_retry_with_new_tips"
-        components={{
-          block: <LegacyStyledText as="p" />,
-        }}
-      />
-    )
-  }
-
-  return (
-    <TwoColTextAndFailedStepNextStep
-      {...props}
-      leftColTitle={t('retry_with_new_tips')}
-      leftColBodyText={buildBodyText()}
-      primaryBtnCopy={t('retry_now')}
-      primaryBtnOnClick={primaryBtnOnClick}
-    />
-  )
 }
