+++ app/src/organisms/ErrorRecoveryFlows/shared/index.ts
@@ -21,5 +21,7 @@ export { SkipStepInfo } from './SkipStepInfo'
 export { GripperIsHoldingLabware } from './GripperIsHoldingLabware'
 export { RecoveryDoorOpenSpecial } from './RecoveryDoorOpenSpecial'
 export { FillWell } from './FillWell'
+export { RetryWithSameTips } from './RetryWithSameTips'
+export { RetryWithNewTips } from './RetryWithNewTips'
 
 export type { RecoveryInterventionModalProps } from './RecoveryInterventionModal'
